\name{ANC data}
\docType{data}
\alias{anc}
\title{Dataset on effect of new ANC method on mortality}
\description{
This dataset contains 755 records of high risk pregnant women 
under a trial on new and old methods of antenatal care 
in clinic A and B. Outcome was perinatal mortality.
}
\usage{data(anc)}
\format{A data frame containing 755 observations and 3 variables.}
\examples{
data(anc)
}
\keyword{datasets}
    