\name{Tooth decay}
\docType{data}
\alias{decay}
\title{Dataset on tooth decay and mutan Streptococcus}
\description{
Relationship between bacteria and presence of any decayed tooth.
     decay = Any decayed tooth  
     strep = CFU of mutan Streptococci in saliva.
}
\usage{data(decay)}
\format{A data frame containing 436 observations and 2 variables.}
\source{Teanpaisan, R., Kintarak, S., Chuncharoen, C., Akkayanont, P. 1995
Mutans Streptococci and dental -caries in schoolchildren 
in Southern Thailand. 
\emph{Community Dentistry and Oral Epidemiology}  \bold{23}: 317-318.}
\examples{
data(decay)
use(decay)
des()
}
\keyword{datasets}
    