\name{Ectopic pregnancy}
\docType{data}
\alias{ectopic}
\title{Dataset of a case-control study looking at history of 
abortion as a risk factor for ectopic pregnancy.}
\description{
This case-control study has one case series and two control groups.

The subjects were recruited based on three types of pregnancy outcome  (outc): 

EP = ectopic pregnancy
IA = women coming for induced abortion
Deli = women admitted for full-term delivery                 

Two exposure variables of interest are:

hia      Previous induced abortion: "never IA" and "ever IA"

gravi    Gravidity (number of pregnacy): "1-2", "3-4" and ">4"                             
}
\usage{data(ectopic)}
\format{A data frame containing 723 observations and 4 variables.}
\keyword{datasets}
    