\name{Hookworm in 1993}
\docType{data}
\alias{hw93}
\title{A study on hookworm prevalence and intensity in 1993}
\description{
A dataset from a cross-sectional survey examining hookworm
infection and shoes wearing

id             Personal identication                 

epg            eggs/g. of feces

age            age in years                             

shoes          Shoe wearing     (yes/no)

intense        Intensity (EPG) group: 0, 1-1,999  and 2000+  

agegr          Age group        


}
\usage{data(hw93)}
\format{A data frame containing 637 observations and 6 variables.}
\keyword{datasets}
    