\name{oswego}
\docType{data}
\alias{oswego}
\title{Dataset on food poisoning}
\description{
This dataset contains 75 records of persons under investigation
for the cause of acute food poisoning after a dinner party.
}
\usage{data(oswego)}
\format{A data frame containing 75 observations and 20 variables.}
\source{EpiInfo package}
\references{
See: \url{http://www.cdc.gov/eis/casestudies/casestudyex.htm}.
}
\keyword{datasets}
    