\name{print cci}
\alias{print.cci}
\title{Print cci results}
\description{Print results for cci and cc commands}
\usage{
\method{print}{cci}(x, ...)
}
\arguments{
       \item{x}{object of class 'cci'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ <cvirasak@medicine.psu.ac.th>}
}
\seealso{'cci'}
\examples{
cci(25, 22, 20, 7)
data(Oswego)
use(Oswego)
cc(ill, chocolate)
}
\keyword{database}                                                                            

