\name{print summ}
\alias{print.summ}
\title{Print 'summ' results}
\description{Print summary of data frame of a variable}
\usage{
\method{print}{summ}(x, ...)
}
\arguments{
       \item{x}{object of class 'summ'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ <cvirasak@medicine.psu.ac.th>}
}
\seealso{'summ'}
\keyword{database}                                                                            

