\name{Air Pollution}
\alias{so2}
\docType{data}
\title{ Air pollution and deaths}
\description{
Deaths in London from 1st-15th Dec 1952
}
\usage{data(so2)}
\format{
  A data frame with 15 observations on the following 4 variables.
  \describe{
    \item{\code{day}}{a numeric vector: the day of Dec 1952}
    \item{\code{deaths}}{a numeric vector: number of deaths}
    \item{\code{smoke}}{a numeric vector: atmospheric smoke (mg/cu.m)}
    \item{\code{SO2}}{a numeric vector: atmospheric sulphur dioxide (parts/million)}
  }
}
\source{
from John F. Osborn, Statistical Exercises in Medical Research, Blackwell 1979
}
\keyword{datasets}
