\name{Voluntary counselling and testing}
\docType{data}
\alias{VCT}
\title{Dataset on attitudes toward VCT}
\description{
This dataset contains information on the records of 315 women working at a tourist destination community.
}
\usage{data(VCT)}
\format{A data frame containing 315 observations and 30 variables with
variable descriptions.}
\examples{
data(VCT)
use(VCT)
des()
}
\keyword{datasets}
    