\name{Blood pressure}
\docType{data}
\alias{BP}
\title{Dataset on blood pressure and determinants}
\description{
This dataset contains 100 records of adults from a small cross-sectional
survey in 2001 on blood pressure and its determinants in a community.
}
\usage{data(BP)}
\format{A data frame containing 100 observations and 6 variables with
variable description.}
\examples{
data(BP)
use(BP)
des()
}
\keyword{datasets}
    