\name{Age at marriage}
\docType{data}
\alias{marryage}
\title{Dataset on age at marriage}
\description{
This dataset contains data on age at first marriage of attendants
at a workshop.

id                         

sex                        

birthyr  year of birth     

educ     level of education : up to bachelor, higher than bachelor 

marital  marital status     : single vs married

maryr    year of marriage  

endyr    year of analysis  
}
\usage{data(marryage)}
\format{A data frame containing 27 observations and 7 variables.}
\keyword{datasets}
    