\name{Sleepiness}
\docType{data}
\alias{sleep3}
\title{Sleepiness among participants in a workshop }
\description{
This dataset contains 15 records of participants in a workshop.
Data include gender, date of birth, sleepiness in the lecture
and in group work, weight and height.
}
\usage{data(sleep3)}
\format{A data frame containing 15 observations and 8 variables.}
\examples{
data(sleep3)
use(sleep3)
des()
}
\keyword{datasets}
    